+++ samplesheets/vue3app/tests/unit/SheetTableHeader.spec.ts
@@ -6,6 +6,7 @@ import { createBootstrap } from 'bootstrap-vue-next/plugins/createBootstrap'
 import SheetTableHeader from '@/components/SheetTableHeader.vue'
 import { useAppStore } from '@/stores/appStore.ts'
 import { type SodarContext } from '@/types.ts'
+import { IRODS_PATH_COPY_MSG, VARIANT_INFO } from '@/constants.ts'
 
 import { copy } from '../testUtils.ts'
 import { sodarContext } from '../data/sodarContext.ts'
@@ -15,9 +16,10 @@ import {
   ASSAY_PLUGIN_TITLE,
   ASSAY_UUID,
   PROJECT_UUID,
+  STUDY_PATH,
   STUDY_PLUGIN_NAME,
   STUDY_PLUGIN_TITLE,
-  STUDY_UUID
+  STUDY_UUID,
 } from '../testConstants.ts'
 
 // Test Data -------------------------------------------------------------------
@@ -38,6 +40,13 @@ config.global.plugins = [createBootstrap()]
 config.global.stubs = {
   IrodsStatsBadge: { template: '<span class="' + statsBadgeClass + '" />'}
 }
+// Mock clipboard (NOTE: has to be done in module root)
+const mockCopy = vi.fn()
+vi.mock('@vueuse/core', async () => {
+  const actual = await vi.importActual('@vueuse/core')
+  return { ...actual, useClipboard: () => ({ copy: mockCopy }) }
+})
+const mockNotifyCb = vi.fn()
 
 // Tests -----------------------------------------------------------------------
 
@@ -52,6 +61,7 @@ describe('SheetTableHeader.vue', () => {
     setActivePinia(createPinia())
     const appStore = useAppStore()
     appStore.currentStudyUuid = STUDY_UUID
+    appStore.notifyCb = mockNotifyCb
     appStore.projectUuid = PROJECT_UUID
     appStore.sodarContext = copy(sodarContext) as SodarContext
   })
@@ -183,6 +193,15 @@ describe('SheetTableHeader.vue', () => {
     expect(wrapper.find('.sodar-ss-study-title-badge').exists()).toBe(true)
     expect(wrapper.find('.' + statsBadgeClass).exists()).toBe(false)
     expect(wrapper.find('.sodar-ss-irods-not-created').exists()).toBe(true)
+  })
+
+  test('copy iRODS path to clipboard on button click', async () => {
+    expect(mockNotifyCb).not.toHaveBeenCalled()
+    const wrapper = mountComponent(studyProps)
+    await wrapper.find('.sodar-ss-irods-copy-btn').trigger('click')
+    expect(mockCopy).toHaveBeenCalledWith(STUDY_PATH)
+    expect(mockNotifyCb).toHaveBeenCalledWith(
+      IRODS_PATH_COPY_MSG, VARIANT_INFO)
   })
 
   // TODO: Test modal opening